+++ hfiles/helpers.h
@@ -9,10 +9,10 @@ using namespace std;
 string str_trim(string str);
 
 typedef struct {
-  uint8_t cmd;
-  uint8_t op1;
-  uint8_t op2;
-  uint8_t op3;
+  uint16_t cmd;
+  uint16_t op1;
+  uint16_t op2;
+  uint16_t op3;
   bool empty;
 } instruction;
